refactor: Improve structure Split ChefCardList into ChefCardItem and SliderSelector
